Deleting the CNAME broke the site.
Update site.baseurl in _config.yml

The site.baseurl must be set for sites built with Jekyll that are
not published at the root of their domain.  This is the default
for GitHub Pages repos http://<user>.github.io/<repo>
or <user custom domain>/<repo>

I forked new-blog-test-5 from new-blog-test-4
Will fixing the baseurl without fixing the url make a site that works
as good as new-blog-test-4?