Fix baseurl for this site
Becasue this site is a clone and both have a site.baseurl and hosted as
sub directories of the same domain the wrong site.baseurl was transferring
 me to the site I forked from.

The whole concept of what is a site and what is a page or group of
pages that is a subset of a site is a bit muddy with the GitHub Pages
model for multiple repos that are not using a Custom domain.

I only caught this becasue I added reporting of the site.url, site.baseurl
and page.url to the index file before I cloned the site for this set of
tests.