Implement menuBar for all Jekyll generated pages.
Implement a drop down navigation list with a link to every page
on the site organised under Post, Page and Static links to the
/Sitemap.html page.  Omit the /About page and site root from the
Pages dropdown list since they are hardcoded as last and first
entries on the menuBar.

This was implemented in _includes/header.html which included in the
_layouts/default.html file affecting all the Jekyll generated pages.

Modify the footer to link to the new site.github_repo value added
to the _config.yml file.  Add Site last updated to the footer.
Add the Top navigation bar Id and class definitions to the stylesheet.